refactor(rag): remove deduplication backwards compatibility and qdrant_vector_size
- Removed legacy deduplication backwards compatibility code handling random UUID chunk IDs.
- Removed qdrant_vector_size config and replaced with embedding_dimension throughout the codebase.
- Updated imports to reflect new module structure for RagService and related classes.
- Simplified QueryCache locking mechanism.
- Cleaned up related tests imports.

This refactor removes deprecated deduplication logic and standardizes vector size configuration to embedding_dimension.
