+++ .env.example
@@ -25,7 +25,7 @@ LOG_TO_FILE=false
 QDRANT_URL=http://localhost:7000
 # QDRANT_API_KEY=your_api_key_here
 QDRANT_COLLECTION=crawlerr_documents
-QDRANT_VECTOR_SIZE=1024
+# QDRANT_VECTOR_SIZE removed - use EMBEDDING_DIMENSION instead
 QDRANT_DISTANCE=cosine
 QDRANT_TIMEOUT=10.0
 QDRANT_RETRY_COUNT=3
