+++ crawler_mcp/core/rag/deduplication.py
@@ -525,60 +525,3 @@ class DeduplicationManager(ABC):
         clean_id = chunk_id.replace("-", "")
         return bool(uuid_pattern.match(clean_id))
 
-    def find_legacy_chunk_by_content(
-        self, existing_chunks: list[dict[str, Any]], content: str
-    ) -> dict[str, Any] | None:
-        """
-        Find a legacy chunk by content hash for backwards compatibility.
-
-        In mixed environments, we need to match content even if the chunk
-        has a random UUID instead of a deterministic ID.
-
-        Args:
-            existing_chunks: List of existing chunks
-            content: Content to match against
-
-        Returns:
-            Matching chunk or None
-        """
-        content_hash = self.generate_content_hash(content)
-
-        for chunk in existing_chunks:
-            # Direct content hash comparison
-            if chunk.get("content_hash") == content_hash:
-                return chunk
-
-            # Fallback: direct content comparison for chunks without hashes
-            if chunk.get("content") == content:
-                return chunk
-
-        return None
-
-    def should_use_backwards_compatibility(
-        self, existing_chunks_map: dict[str, str]
-    ) -> bool:
-        """
-        Determine if backwards compatibility mode should be used.
-
-        This checks if the existing chunks contain random UUIDs, indicating
-        a mixed environment that needs backwards compatibility handling.
-
-        Args:
-            existing_chunks_map: Map of chunk_id -> content_hash
-
-        Returns:
-            True if backwards compatibility is needed
-        """
-        if not existing_chunks_map:
-            return False
-
-        # Sample a few chunk IDs to determine if they're random UUIDs
-        sample_size = min(5, len(existing_chunks_map))
-        sample_ids = list(existing_chunks_map.keys())[:sample_size]
-
-        random_uuid_count = sum(
-            1 for chunk_id in sample_ids if self.is_random_uuid(chunk_id)
-        )
-
-        # If more than half are random UUIDs, use backwards compatibility
-        return random_uuid_count > (sample_size // 2)
